// src/app/models/luminaria.model.ts
export class Luminaria {
    observaciones?: string;
    id_luminaria?: string;
    punto_luz?: string;
    vial?: string;
    numero?: number;
    centro_mando?: string;
    circuito?: string;
    operativa?: string;
    altura?: number;
    tipo_soporte?: string;
    marca_soporte?: string;
    modelo_soporte?: string;
    estado_soporte?: string;
    tipo_luminaria?: string;
    marca_luminaria?: string;
    modelo_luminaria?: string;
    estado_luminaria?: string;
    tipo_lampara?: string;
    marca_lampara?: string;
    modelo_lampara?: string;
    estado_lampara?: string;
    cantidad_lamparas?: number;
    potencia?: number;
    equipo_auxiliar?: string;
    situacion_equipo_auxiliar?: string;
    alta?: false;
    usuario_alta?: string;
    fecha_alta?: string;
    modificado?: false;
    usuario_modificado?: string;
    fecha_modificado?: string;
    herramienta?: string;
    numero_registro?: number;
    distancia_eje?: string;
    lado_via?: string;
    longitud_brazo?: string;
    info_completa?: boolean = false;
}